+++ package.json
@@ -19,7 +19,7 @@
   "scripts": {
     "start:fullstack": "cross-env REACT_APP_WHOAMI=fullstack react-scripts start",
     "fe": "cross-env REACT_APP_WHOAMI=frontend react-scripts start --no-sourcemap",
-    "build": "react-scripts build",
+    "build": "cross-env REACT_APP_WHOAMI=frontend react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
