+++ src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ToDoPage from './ToDoPage';
+import ToDoPage from "pages/ToDoPage";
 
 function App() {
     return (
