+++ src/assets/scss/App/main.scss
@@ -21,4 +21,5 @@
 
 @import "todo-creation";
 @import "todo-list";
+@import "todo-item";
 @import "todo-toolbar";
