+++ src/service/types.ts
@@ -1,6 +1,9 @@
 import {Todo} from '../models/todo';
 
 export abstract class IAPI {
-    abstract getTodos() : Promise<Array<Todo>>
-    abstract createTodo(content: string) : Promise<Todo>
+    abstract getTodos(): Promise<Array<Todo>>
+
+    abstract updateTodos(todos: Array<Todo>): Promise<any>
+
+    abstract createTodo(content: string): Promise<Todo>
 }
