+++ src/utils/index.ts
@@ -1,5 +1,7 @@
 import {Todo, TodoStatus} from '../models/todo';
 
+const LOCALSTORAGE_NAMESPACE = "todos";
+
 export function isTodoCompleted(todo: Todo): boolean {
     return todo.status === TodoStatus.COMPLETED;
 }
@@ -8,7 +10,7 @@ export function isTodoActive(todo: Todo): boolean {
     return todo.status === TodoStatus.ACTIVE;
 }
 
-export function localStore(namespace: string, data?: any) {
+export function localStoreData(data?: any, namespace: string = LOCALSTORAGE_NAMESPACE) {
     if (data) {
         return localStorage.setItem(namespace, JSON.stringify(data));
     }
